+++ src/components/ProductCard/ProductCard.props.ts
@@ -1,0 +1,15 @@
+export type TProductSize = {
+    //типизация размеров
+    title: string;
+    number: number;
+};
+
+export type TProductData = {
+    imageURL: string;
+    title: string;
+    isFavorite: boolean;
+    size: TProductSize[];
+    price: number;
+    oldPrice: number;
+    currency: string;
+};
